refactor: create tabs interface for push and xmtp (#3)
* create tabbed interface for push and xmtp

* Changes
1> Separated all push and xmtp into separate folders
2> Changed provider query into /xmtp or /push
3> Changed prev useMessageStore to useXmtpStore in xmtp-message.ts

---------
