+++ apps/web/src/components/Messages/Xmtp/index.ts
@@ -1,0 +1,6 @@
+export * from './Composer';
+export * from './Message';
+export * from './MessageHeader';
+export * from './MessageIcon';
+export * from './Preview';
+export * from './XMTPPreview';
