--- apps/web/src/pages/messages/[...conversationKey].tsx
@@ -1,3 +1,0 @@
-import Message from '@components/Messages/Message';
-
-export default Message;
