--- apps/web/src/pages/pushMessage/index.tsx
@@ -1,3 +1,0 @@
-import PushMessage from '@components/PushMessage';
-
-export default PushMessage;
